Add files via upload
We have a list of measurements for the lengths of several metal objects in feet for various
temperatures in Reaumur degrees. Since we are not familiar with nonstandard metrics, we
want to convert the measurements to standard metrics, i.e., we want lengths to be expressed
in centimeters and the temperatures in Celsius degrees.