Add files via upload
Our program should take two lines of input. The first line has the information for the coordinates of p1 and the second line has the information for the coordinates of p2 . Each lineconsists of two numbers that are the x and y coordinates of the respective point. Note that a coordinate of a point is not necessarily an integer.
We wants to write a Java program which computes the distance between two given points on two-dimensional Euclidean space with respect to the distance metrics below.
ManhattanDistance(p 1 , p 2 ) : |x 1 − x 2 | + |y 1 − y 2 |
Euclidean Distance(p 1 , p 2 ) : (x 1 − x 2 )^2 + (y 1 − y 2 )^2
Cosine Distance(p 1 , p 2 ) : 1 − ((x1 ∗ x2) + (y 1 ∗ y 2))/ √¯ x^2 + y^2 ∗ √¯x^2 + y^2