Add files via upload
Write two Java programs that helps users to encrypt and decrypt messages with the Ceasar
cipher. Your first program should prompt the user to enter a message which is referred to as
“plaintext" in cryptography. Your program should echo the input and report the encrypted
message, which is referred to as “ciphertext" in cryptography. Your second program should
reverse this process, i.e., it should prompt the user to provide a ciphertext and then report the
corresponding plaintext.

Before starting to write the program, make research on the Ceasar cipher. There are different
variations of the Caesar cipher, in this homework, you will use the one that uses a right shift
of 3 to encrypt a message, i.e., a → d , b → e, . . . w → z, x → a, y → b, z → c. The program that
does the decryption should reverse the process, i.e., decryption is via a left shift of 3, which
is the same as right shift of 23.