Add files via upload
Write a Java program that does the following. Your program should prompt the user to pro-
vide the name, surname and the year of birth of a famous person. It should then prompt the
user the second time to provide a quote of this famous person. It should then prompt the
user the third time to provide the year of the quote. The program should report the quote,
and the age of the person at that time.