Add files via upload
We have a gambler that came to casino with $n in his pocket. He wants to make his money
$m for some m > n, by playing as many rounds as needed of a particular game. In each
round of the game, the gambler bets $1, and wins $1 with some probability p or loses $1 with
probability 1 − p. Notice that at the end of a round, the gambler has one more dollar than the
end of the previous round with probability p, and he has one less dollar than the end of the
previous round with probability 1− p. The gambler plays this game until one of the following
two things happens: (i) his money reaches the goal he set before coming to the casino, i.e.,
his money becomes $m, which we refer to as a win for him, (ii) his money reaches $0 and he
is not permitted to play any more rounds, and we refer to this situation as a lose for him. As
you will see, she has a very slim chance of winning if m is large for p < 0.5.
As simulations involve randomness, you will simulate this game several times and count how
many times the gambler won and lost. Your program must take the following as input and
output how many times the gambler won and lost.
• n : the initial money of the gambler,
• m : the goal of the gambler,
• p : the probability of winning at each round, and
• the number of times the game is to be simulated. Note that one simulation of the game
consists of several rounds.