Add files via upload
You have just purchased a new laptop for your Bil113 course. Laptop costs $1, 500 on the
following credit plan: no down payment, an interest rate of 18% per year (and hence 1.5%
per month), and monthly payments of $50. The monthly payment of $50 is used to pay the
interest, and whatever is left is used to pay part of the remaining debt. Hence, the first month
you pay 1.5% of $1, 500 in interest. That is $22.5 in interest. So, the remaining $27.50 is de-
ducted from your debt, which leaves you with a debt of $1472.50. The next month, you pay
interest of 1.5% of $1472.50, which is $22, 0875. Hence, you can deduct $27.9125 (which is
$50–$22, 0875) from the amount you owe. This goes on until you have no debt.
In this task you will write a program that tells you how many months it will take you to pay
off a loan with the given interest under equal payments each month. Also, you will calculate
the total payment paid. Use a loop to calculate the amount of interest and the size of the debt
after each month. (Your final program need not output the monthly amount of interest paid
and remaining debt, but you may want to write a preliminary version of the program that
does output these values.) Use a variable to count the number of loop iterations and hence
the number of months until the debt is zero. You may want to use other variables as well.
The last payment may be less than the given monthly payment if the debt is small, but do not
forget the interest! If you owe $50, your monthly payment of $50 will not pay off your debt,
although it will come close. One month’s interest on $50 is only 75 cents.